AddTool uses label as template for unique label
When AddTool adds a component with an existing label, a suffix is added
if needed, using the existing label as a template, to ensure the
resulting component label is unique within the circuit.

Also, XmlReader no longer discards labels within tool and library
components, so these labels can be used as templates for AddTool.