Remove slf4j Logger dependency
Eliminate all uses of the slf4j logger classes, and remove the jar
dependency. This logger was used very inconsistently, and in the places
where it was used, it was not used in any particularly useful way. In
amny cases, it was used simply to print fatal errors before dying.