fastpath for CircuitState.setValueByWire()
Use a 2d array of Values (large enough for a typical max canvas size) as
a fastpath instead of the usual hashmap from points to values. This
helped a bit in some circuits, but not in others. The real solution to
speeding up simulation is to re-think the Circuitstate / WireBundle /
CircuitWires implementation so it is not point-based. Maybe
netlist-based instad?