linked list queue for pending event set
This implementation is about as simple as it gets, just a doubly-linked
list. Yet the performance seems even higher than the splay tree. We
insert using a linear scan from the end of the list. The common case by
far is to insert at the tail of the list, so this is usually efficient.

cpuM2.zip/cpuhci.circ now runs at about 4.0-4.5kHz, up from 3.2kHz.