Fix backspace vs delete issues on Mac
On Mac, backspace and delete are reversed compared to the rest of thw
world. The key labeled "delete" on the keyboard generates a backspace
event, so most code works fine. However, setting VK_DELETE as a menu
accelerator results in a menu with a backspace-like icon on Mac, instead
of the word "DELETE" or "DEL" as on other systems, and it is harder for
mac users to generate VK_DELETE event (since they need to find the mac's
backspace key to do so, or use Fn+Delete).

Workaround: Make the menu accelerator by VK_BACK_SPACE on mac, which
still appears as the backspace icon, but can be generated more easily by
the user. Also make any code that handles VK_DELETE also handle
VK_BACK_SPACE, and vice versa.