Version bump, and fix quad-input wide gate layout
Certain primitive gates (AND, OR, NAND, NOR, XOR, XNOR, Even Parity, and
Odd Parity) that were configured with quad-input and "Wide" appearance
had a bad layout for their inputs. The input pin layout is now fixed:
this breaks backwards compatibility with files saved in previous
versions. A warning is added to this effect upon file open.

Version bumped to 4.0.2 HC.