region-select can select partial wires
First attempt, not yet fully working.

When using the selection tool (or multi-select tool) and selecting a
region of a canvas, the user can select parts of wires then
cut/copy/delete/move them.

Pros:
- You can either move a component alone and it will reconnect the pins
by repairing wires if possible (old behavior). Or you can now also
select a little bit of the wires nearby and it will not reconnect the
pins but instead just move those portions of the wires.
- You can easily delete regions of the canvas by selecting then
deleting. Any component within the region will be deleted (old
behavior), but now any wires passing through the region will also get
partially removed.
- You can easily copy-paste sections of wires.

Cons and remaining issues:
- Just selecting a region counts as an action and changes the circuit.
This is because it it splits the wires and "lifts" the affected portion
out of the circuit, as if it had just been pasted and is floating but
not yet anchored.
- Just selecting a region can affect the simulation in progress, because
the wire portions are removed, and the original connections eliminated,
at the time of selection. Also, the addition of the leftover wire
portions can also cause new connections to get created where there used
to be only a wire-crossing, and again this happens at the time of
selection, which is awkward.
- The undo behavior is a bit wierd, because of the select counting as an
independent action.
- Selecting densly-wired regions with lots of crossings and
intersections is tricky and confusing. The undo/unselect behavior here
is particularly bad.
- There is a bug in the wire bounds/intersection code somewhere, such
that certain wire selections don't work (e.g. selecting just long thin
region on an existing wire doesn't always work).